feat(ui): Refactor home screen to use AppBar and sections
This commit refactors the home screen to use a standard AppBar
instead of a SliverAppBar, improving layout consistency. It also
introduces separate widgets for section titles and moves the
`EventSearchDelegate` to its own file for better organization.
Additionally, the program info card was removed, and workshops
are displayed in a card with more details.